backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
mysqlclient==2.2.8
python-dotenv==1.0.1
requests==2.32.3
edge-tts==6.1.12
gTTS==2.5.3
deep-translator==1.11.4
Pillow==12.2.0
